Remove the compliance history API tests
The tech preview feature is being removed so the tests are no longer
needed.

Relates:
https://issues.redhat.com/browse/ACM-15291
https://issues.redhat.com/browse/ACM-13208
