--- test/resources/compliance_history/cert-policy.yaml
@@ -1,53 +1,0 @@
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: cert-policy
-  annotations:
-    policy.open-cluster.management.io/standards: NIST-CSF
-    policy.open-cluster.management.io/categories: PR.PT Protective Technology
-    policy.open-cluster.management.io/controls: PR.PT-3 Least Functionality
-spec:
-  remediationAction: inform
-  disabled: false
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: CertificatePolicy
-        metadata:
-          name: cert-policy
-        spec:
-          remediationAction: inform
-          namespaceSelector:
-            exclude: ["kube-*"]
-            include: ["ch-cert-policy-test-ns"]
-          severity: low
-          minimumDuration: 30h
-          allowedSANPattern: ".*.test.com"
----
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: compliance-history-test
-  namespace: policy-test
-spec:
-  predicates:
-    - requiredClusterSelector:
-        labelSelector:
-          matchLabels:
-            feature.open-cluster-management.io/addon-cert-policy-controller: "available"
-            feature.open-cluster-management.io/addon-config-policy-controller: "available"
-            feature.open-cluster-management.io/addon-governance-policy-framework: "available"
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: compliance-history-test
-  namespace: policy-test
-placementRef:
-  name: compliance-history-test
-  apiGroup: cluster.open-cluster-management.io
-  kind: Placement
-subjects:
-  - name: cert-policy
-    apiGroup: policy.open-cluster-management.io
-    kind: Policy
