--- test/resources/compliance_history/operator-policy-invalid.yaml
@@ -1,54 +1,0 @@
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: op-compliance-api
-  annotations:
-    policy.open-cluster-management.io/description: dog
-spec:
-  disabled: false
-  remediationAction: inform
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1beta1
-        kind: OperatorPolicy
-        metadata:
-          name: test-operator-policy
-        spec:
-          remediationAction: inform
-          severity: medium
-          complianceType: musthave
-          subscription:
-            channel: stable-3.8
-            name: quay-operator
-            namespace: ch-operator-policy-test-ns
-            # correct one: redhat-operators
-            source: invalid
-            sourceNamespace: openshift-marketplace
-          upgradeApproval: Automatic
----
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: compliance-history-test
-  namespace: policy-test
-spec:
-  predicates:
-    - requiredClusterSelector:
-        labelSelector:
-          matchLabels:
-            feature.open-cluster-management.io/addon-config-policy-controller: "available"
-            feature.open-cluster-management.io/addon-governance-policy-framework: "available"
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: compliance-history-test
-  namespace: policy-test
-placementRef:
-  name: compliance-history-test
-  apiGroup: cluster.open-cluster-management.io
-  kind: Placement
-subjects:
-  - name: op-compliance-api
-    apiGroup: policy.open-cluster-management.io
-    kind: Policy
