--- test/resources/compliance_history/operator-prereq.yaml
@@ -1,55 +1,0 @@
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: ch-operator-prereq-policy
-  namespace: policy-test
-  annotations:
-    policy.open-cluster-management.io/categories: CM Configuration Management
-    policy.open-cluster-management.io/controls: CM-2 Baseline Configuration
-    policy.open-cluster-management.io/standards: NIST SP 800-53
-spec:
-  remediationAction: enforce
-  disabled: false
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: ch-operator-prereq-configpolicy
-        spec:
-          pruneObjectBehavior: DeleteAll
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: Namespace
-                metadata:
-                  name: ch-operator-policy-test-ns
-          severity: critical
----
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: ch-operator-prereq
-  namespace: policy-test
-spec:
-  predicates:
-    - requiredClusterSelector:
-        labelSelector:
-          matchLabels:
-            feature.open-cluster-management.io/addon-config-policy-controller: "available"
-            feature.open-cluster-management.io/addon-governance-policy-framework: "available"
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: ch-operator-prereq
-  namespace: policy-test
-placementRef:
-  name: ch-operator-prereq
-  apiGroup: cluster.open-cluster-management.io
-  kind: Placement
-subjects:
-  - name: ch-operator-prereq-policy
-    apiGroup: policy.open-cluster-management.io
-    kind: Policy
