--- test/resources/compliance_history/policy-gk-prereq.yaml
@@ -1,62 +1,0 @@
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: compliance-api-gk-prereq
-  namespace: policy-test
-  annotations:
-    policy.open-cluster-management.io/categories: CM Configuration Management
-    policy.open-cluster-management.io/controls: CM-2 Baseline Configuration
-    policy.open-cluster-management.io/standards: NIST SP 800-53
-spec:
-  remediationAction: enforce
-  disabled: false
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: compliance-api-configure-namespace-configmap
-        spec:
-          pruneObjectBehavior: DeleteAll
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: Namespace
-                metadata:
-                  name: compliance-api-test
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: ConfigMap
-                metadata:
-                  name: compliance-api-test
-                  namespace: compliance-api-test
-          severity: critical
----
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: compliance-api-gk-prereq
-  namespace: policy-test
-spec:
-  predicates:
-    - requiredClusterSelector:
-        labelSelector:
-          matchLabels:
-            feature.open-cluster-management.io/addon-config-policy-controller: "available"
-            feature.open-cluster-management.io/addon-governance-policy-framework: "available"
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: compliance-api-gk-prereq
-  namespace: policy-test
-placementRef:
-  name: compliance-api-gk-prereq
-  apiGroup: cluster.open-cluster-management.io
-  kind: Placement
-subjects:
-  - name: compliance-api-gk-prereq
-    apiGroup: policy.open-cluster-management.io
-    kind: Policy
