--- test/resources/compliance_history/policy-install-gk.yaml
@@ -1,93 +1,0 @@
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: compliance-api-install-gk
-  namespace: policy-test
-spec:
-  disabled: false
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1beta1
-        kind: OperatorPolicy
-        metadata:
-          name: compliance-api-install-gk
-        spec:
-          remediationAction: enforce
-          severity: critical
-          complianceType: musthave
-          subscription:
-            channel: stable
-            name: gatekeeper-operator-product
-            namespace: openshift-operators
-            source: redhat-operators
-            sourceNamespace: openshift-marketplace
-          upgradeApproval: Automatic
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: compliance-api-configure-gk
-        spec:
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: operator.gatekeeper.sh/v1alpha1
-                kind: Gatekeeper
-                metadata:
-                  name: gatekeeper
-                spec:
-                  validatingWebhook: Disabled
-                  mutatingWebhook: Disabled
-                  audit:
-                    auditInterval: 10s
-                    resources:
-                      limits:
-                        memory: 1Gi
-          remediationAction: enforce
-          severity: critical
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: gatekeeper-status
-        spec:
-          remediationAction: inform
-          severity: high
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: Pod
-                metadata:
-                  namespace: openshift-gatekeeper-system
-                  labels:
-                    control-plane: audit-controller
-                status:
-                  phase: Running
----
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: compliance-api-install-gk
-  namespace: policy-test
-spec:
-  predicates:
-    - requiredClusterSelector:
-        labelSelector:
-          matchLabels:
-            feature.open-cluster-management.io/addon-config-policy-controller: "available"
-            feature.open-cluster-management.io/addon-governance-policy-framework: "available"
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: compliance-api-install-gk
-  namespace: policy-test
-placementRef:
-  name: compliance-api-install-gk
-  apiGroup: cluster.open-cluster-management.io
-  kind: Placement
-subjects:
-  - name: compliance-api-install-gk
-    apiGroup: policy.open-cluster-management.io
-    kind: Policy
