--- test/resources/compliance_history/policy.yaml
@@ -1,69 +1,0 @@
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: compliance-api-configpolicy
-  namespace: policy-test
-  annotations:
-    policy.open-cluster-management.io/categories: CM Configuration Management
-    policy.open-cluster-management.io/controls: CM-2 Baseline Configuration
-    policy.open-cluster-management.io/standards: NIST SP 800-53
-spec:
-  disabled: false
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: default-namespace-must-exist
-        spec:
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: Namespace
-                metadata:
-                  name: default
-          remediationAction: inform
-          severity: low
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: does-not-exist-namespace-must-exist
-        spec:
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: Namespace
-                metadata:
-                  name: does-not-exist
-          remediationAction: inform
-          severity: low
----
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: compliance-api-configpolicy
-  namespace: policy-test
-spec:
-  predicates:
-    - requiredClusterSelector:
-        labelSelector:
-          matchLabels:
-            feature.open-cluster-management.io/addon-config-policy-controller: "available"
-            feature.open-cluster-management.io/addon-governance-policy-framework: "available"
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: compliance-api-configpolicy
-  namespace: policy-test
-placementRef:
-  name: compliance-api-configpolicy
-  apiGroup: cluster.open-cluster-management.io
-  kind: Placement
-subjects:
-  - name: compliance-api-configpolicy
-    apiGroup: policy.open-cluster-management.io
-    kind: Policy
