--- test/resources/compliance_history/service_account.yaml
@@ -1,42 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: compliance-history-user
-  namespace: default
----
-apiVersion: v1
-kind: Secret
-type: kubernetes.io/service-account-token
-metadata:
-  name: compliance-history-user
-  namespace: default
-  annotations:
-    kubernetes.io/service-account.name: compliance-history-user
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  creationTimestamp: null
-  name: compliance-history-user
-rules:
-- apiGroups:
-  - cluster.open-cluster-management.io
-  resources:
-  - managedclusters
-  resourceNames:
-  - "*"
-  verbs:
-  - get
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: compliance-history-user
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: compliance-history-user
-subjects:
-- kind: ServiceAccount
-  name: compliance-history-user
-  namespace: default
